move nslsii to src
