+++ src/nslsii/common/ipynb/animation.py
