+++ src/nslsii/detectors/QEPro.py
