+++ src/nslsii/detectors/maia.py
