+++ src/nslsii/detectors/trigger_mixins.py
