+++ src/nslsii/detectors/utils.py
