+++ src/nslsii/detectors/webcam.py
