+++ src/nslsii/detectors/xspress3.py
