+++ src/nslsii/detectors/zebra.py
